\name{guidedPLS-package}
\alias{guidedPLS-package}
\docType{package}
\title{
\packageTitle{guidedPLS}
}
\description{
\packageDescription{guidedPLS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{guidedPLS}
\packageIndices{guidedPLS}
}
\author{
\packageAuthor{guidedPLS}

Maintainer: \packageMaintainer{guidedPLS}
}
\references{
Le Cao, et al. (2008). A Sparse PLS for Variable Selection when Integrating Omics Data. \emph{Statistical Applications in Genetics and Molecular Biology}, 7(1)

Reese S E, et al. (2013). A new statistic for identifying batch effects in high-throughput genomic data that uses guided principal component analysis. \emph{Bioinformatics}, 29(22), 2877-2883
}

\keyword{ package }
\seealso{\code{\link{toyModel}},\code{\link{PLSSVD}},\code{\link{sPLSDA}},\code{\link{guidedPLS}}}
\examples{
ls("package:guidedPLS")
}